remove duplicated get_children in both util.py, rename interface/util -> interface/helper
